+++ vitest.config.ts
@@ -1,3 +1,4 @@
+
 import { defineConfig } from 'vitest/config'
 import { playwright } from '@vitest/browser-playwright'
 import { getDirname } from './src/node/getDirname'
@@ -5,11 +6,27 @@ import path from 'node:path'
 
 const __dirName = getDirname(import.meta.url)
 
+const alias = [
+  {
+    find: new RegExp('@common'), replacement: path.resolve(__dirName, 'src', 'common'),
+  },
+  {
+    find: new RegExp('@browser'), replacement: path.resolve(__dirName, 'src', 'browser'),
+  },
+  {
+    find: new RegExp('@node'), replacement: path.resolve(__dirName, 'src', 'node'),
+  },
+  {
+    find: new RegExp('@constants'), replacement: path.resolve(__dirName, 'src', 'constants'),
+  }
+]
+
 export default defineConfig({
   test: {
     projects: [
       {
         test: {
+          alias,
           include: [
             './src/common/**/*.test.ts',
           ],
@@ -19,6 +36,7 @@ export default defineConfig({
       },
       {
         test: {
+          alias,
           include: [
             './src/node/**/*.test.ts',
           ],
@@ -29,14 +47,7 @@ export default defineConfig({
       {
 
         test: {
-          alias: [
-            {
-              find: new RegExp('@common'), replacement: path.resolve(__dirname, 'src', 'common'),
-            },
-            {
-              find: new RegExp('@browser'), replacement: path.resolve(__dirname, 'src', 'browser'),
-            }
-          ],
+          alias,
           include: [
             './src/browser/**/*.test.ts',
           ],